requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database
sqlalchemy>=2.0.0

# File uploads
python-multipart>=0.0.9

# Rate limiting
slowapi>=0.1.9

# ML
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0

# RAG - embeddings + vector store
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# OpenAI (optional - only needed if LLM_PROVIDER=openai)
openai>=1.30.0

# Env
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
